% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdocumentation.R
\name{check_package}
\alias{check_package}
\title{Check if a package is installed for the user.}
\usage{
check_package(mypkg)
}
\arguments{
\item{mypkg}{Name of the package}
}
\value{
FALSE if the package is not installed, otherwise the versionnumber of the package
}
\description{
Check if a package is installed for the user.
}
\examples{
check_package("Rdocumentation")
check_package("utils")
}

