% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdocumentation.R
\name{makeDefault}
\alias{makeDefault}
\title{Makes the Rdocumentation package the default help-package.}
\usage{
makeDefault()
}
\description{
Makes the Rdocumentation package the default help-package.
}

