\name{RDocumentation-package}
\alias{RDocumentation-package}
\alias{Rdocumentation}
\alias{RDocumentation}
\docType{package}
\title{
Integrate R with RDocumentation.org
}
\description{
Enhance the search/help functionality in R with RDocumentation.org
}
\details{
\tabular{ll}{
Package: \tab RDocumentation\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-08-09\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jonathan Cornelissen (for RDocumentation.org)
Maintainer: Jonathan Cornelissen <Jonathan@datamind.org>
}
\keyword{ package }
\seealso{
\code{\link[utils]{help}}
}
\examples{
help(mean,base)
?centre
??file.download
}