% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_help.R
\name{view_help}
\alias{view_help}
\title{View the help}
\usage{
view_help(body, arg1, arg2)
}
\arguments{
\item{body}{body of the POST request to RDocumentation}

\item{arg1}{Same as arg1 of utils function}

\item{arg2}{Same as arg2 of utils function}
}
\description{
View the help
}
\details{
Leverage https://www.rdocumentation.org/packages/tools/versions/3.3.1/topics/startDynamicHelp to render html page into the help pane
As the page are rendered by the internal RStudio Server, it tricks RStudio into thinking that page is from the same origin as the other elements
in RStudio IDE, which allows us to communicate from the javascript to the /rpc/... endpoint
}

