% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDocumentation.R
\name{check_package}
\alias{check_package}
\title{Check if a package is installed for the user.}
\usage{
check_package(mypkg, version)
}
\arguments{
\item{mypkg}{Name of the package}

\item{version}{the latest version to be checked}
}
\value{
FALSE if the package is not installed, otherwise the versionnumber of the package
}
\description{
Check if a package is installed for the user.
}
\examples{
check_package("RDocumentation","0.2")
check_package("utils","3.3.1")
}

