% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package}
\alias{check_package}
\title{Check if a package is installed for the user.}
\usage{
check_package(pkg, version)
}
\arguments{
\item{pkg}{Name of the package}

\item{version}{the latest version to be checked}
}
\value{
1 if the package is not installed; -1 if the package is not up to date; 0 if the package if the package is installed and up to date.
}
\description{
Check if a package is installed for the user.
}
\examples{
\dontrun{
check_package("RDocumentation","0.2")
check_package("utils","3.3.1")
}

}

