% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{defaults}
\alias{defaults}
\alias{disable_autoload}
\alias{disable_rdocs}
\alias{enable_autoload}
\alias{enable_rdocs}
\alias{makeDefault}
\title{Settings}
\usage{
enable_autoload()

makeDefault()

disable_autoload()

enable_rdocs()

disable_rdocs()
}
\description{
Automatically load RDocumentation and make it active, so that it overrides the default 
documentation?
}
\details{
When you load in the RDocumentation package, you'll be prompted to autoload 
the package whenever you start R. When RDocumentation is loaded, it will automatically
override the default help functionality with documentation from RDocumentation.org.

The autoloading is done by modyfing the .RProfile file that is in your working directory.

With \code{enable_autoload} and \code{disable_autoload} you can update the
.RProfile file to enable or disable the automatic loading of RDocumentation
when you start R. With \code{enable_rdocs} and \code{disable_rdocs} you can enable or 
disable the overriding of the default help viewer.
}

