% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdocumentation.R
\name{help}
\alias{?}
\alias{help}
\title{Documentation on Rdocumentation or via the normal help system if offline}
\usage{
help(topic, package = NULL, lib.loc = NULL,
  verbose = getOption("verbose"),
  try.all.packages = getOption("help.try.all.packages"),
  help_type = getOption("help_type"))

"?"(e1, e2)
}
\arguments{
\item{topic}{usually, a name or character string specifying the topic for which help is sought.
A character string (enclosed in explicit single or double quotes) is always taken as naming a \code{topic}.
If the value of \code{topic} is a length-one character vector the topic is taken to be the value of the only element.
Otherwise topic must be a name or a reserved word (if syntactically valid) or character string. See ‘Details’ for what happens if this is omitted.}

\item{package}{a name or character vector giving the packages to look into for documentation, or \code{NULL}.
By default, all packages whose namespaces are loaded are used. To avoid a name being deparsed use e.g. \code{(pkg_ref)} (see the examples).}

\item{lib.loc}{a character vector of directory names of R libraries, or \code{NULL}. The default value of \code{NULL} corresponds to all libraries currently known.
If the default is used, the loaded packages are searched before the libraries. This is not used for HTML help (see ‘Details’).
When no local matches are found, Rdocumentation will ignore local libraries and search in the online database.}

\item{verbose}{logical; if \code{TRUE}, the file name is reported.}

\item{try.all.packages}{logical; see \code{Note}.}

\item{help_type}{character string: the type of help required. Possible values are code{"text"}, \code{"html"} and code{"pdf"}. Case is ignored, and partial matching is allowed.}

\item{topic}{Usually, a name or character string specifying the topic for which help is sought. Alternatively, a function call to ask for documentation on a corresponding S4 method: see the section on S4 method documentation. The calls pkg::topic and pkg:::topic are treated specially, and look for help on topic in package pkg.}

\item{type}{the special type of documentation to use for this topic; for example, if the type is class, documentation is provided for the class with name topic. See the Section ‘S4 Method Documentation’ for the uses of type to get help on formal methods, including methods?function and method?call.

This is a shortcut to help and uses its default type of help.
Some topics need to be quoted (by backticks) or given as a character string. There include those which cannot syntactically appear on their own such as unary and binary operators, function and control-flow reserved words (including if, else for, in, repeat, while, break and next. The other reserved words can be used as if they were names, for example TRUE, NA and Inf.}
}
\description{
\code{help} contacts Rdocumentation for help given aliases and packages

These functions provide access to Rdocumentation. Documentation on a topic with name name (typically, an R object or a data set) can be displayed by either help("name") or ?name.
}
\examples{
help(package=Rdocumentation)
help(centre,ftsa)

?"for"                  # but quotes/backticks are needed
?`+`

?women                  # information about data set "women"

## Not run: 
# require(methods)
# ## define a S4 generic function and some methods
# combo <- function(x, y) c(x, y)
# setGeneric("combo")
# setMethod("combo", c("numeric", "numeric"), function(x, y) x+y)
# 
# ## assume we have written some documentation
# ## for combo, and its methods ....
# 
# ?combo  # produces the function documentation
# 
# methods?combo  # looks for the overall methods documentation
# 
# method?combo("numeric", "numeric")  # documentation for the method above
# 
# ?combo(1:10, rnorm(10))  # ... the same method, selected according to
#                          # the arguments (one integer, the other numeric)
# 
# ?combo(1:10, letters)    # documentation for the default method
# ## End(Not run)
}
\keyword{documentation}

