% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrides.R
\name{help.search}
\alias{help.search}
\title{Search the Help System on RDocumentation or local if offline}
\usage{
help.search(...)
}
\arguments{
\item{pattern}{a character string to be matched in the specified fields. If this is given, the arguments \code{apropos}, \code{keyword}, and \code{whatis} are ignored.}

\item{fields}{a character vector specifying the fields of the help database to be searched. The entries must be abbreviations of \code{"name"}, \code{"title"}, \code{"alias"}, \code{"concept"}, and \code{"keyword"},
corresponding to the help page's (file) name, its title, the topics and concepts it provides documentation for, and the keywords it can be classified to.
See below for details and how vignettes and demos are searched.}

\item{apropos}{a character string to be matched in the help page topics and title.}

\item{keyword}{a character string to be matched in the help page ‘keywords’. ‘Keywords’ are really categories: the standard categories are listed in file ‘R.home("doc")/KEYWORDS’
(see also the example) and some package writers have defined their own. If \code{keyword} is specified, \code{agrep} defaults to \code{FALSE}.}

\item{whatis}{a character string to be matched in the help page topics.}

\item{ignore.case}{a logical. If \code{TRUE}, case is ignored during matching; if \code{FALSE}, pattern matching is case sensitive. If no local matches are found, the online search is used, which is always non case-sensitive.}

\item{package}{a character vector with the names of packages to search through, or \code{NULL} in which case all available packages in the library trees specified by \code{lib.loc} are searched.
if no matches are found, the RDocumentation.org database is searched.}

\item{lib.loc}{a character vector describing the location of R library trees to search through, or \code{NULL}. The default value of \code{NULL} corresponds to all libraries currently known. If no matches are found in the given library,
the online search of RDocumentation is used.}

\item{help.db}{a character string giving the file path to a previously built and saved help database, or \code{NULL}.}

\item{verbose}{logical; if \code{TRUE}, the search process is traced. Integer values are also accepted, with \code{TRUE} being equivalent to \code{2},
' and \code{1} being less verbose. On Windows a progress bar is shown during rebuilding, and on Unix a heartbeat is shown for \code{verbose = 1} and a package-by-package list for \code{verbose >= 2}.}

\item{rebuild}{a logical indicating whether the help database should be rebuilt. This will be done automatically if \code{lib.loc} or the search path is changed, 
or if \code{package} is used and a value is not found.}

\item{agrep}{if \code{NULL} (the default unless \code{keyword} is used) and the character string to be matched consists of alphanumeric characters,
whitespace or a dash only, approximate (fuzzy) matching via \code{\link[base]{agrep}} is used unless the string has fewer than 5 characters; 
otherwise, it is taken to contain a regular expression (\code{\link[base]{regex}}) to be matched via \code{\link[base]{grep}}. If \code{FALSE},
approximate matching is not used. Otherwise, one can give a numeric or a list specifying the maximal distance for the approximate match, see argument \code{max.distance} in the documentation for agrep.}

\item{use_UTF8}{logical: should results be given in UTF-8 encoding? Also changes the meaning of regexps in \code{agrep} to be Perl regexps. This does not have effect when the online database is used.}

\item{types}{a character vector listing the types of documentation to search. The entries must be abbreviations of \code{"vignette"}, \code{"help"} or \code{"demo"}. Results will be presented in the order specified.}
}
\description{
Allows for searching the help system for documentation matching a given character string in the (file) name, alias, title, concept or keyword entries (or any combination thereof),
using either fuzzy matching(\code{\link[base]{agrep}}) or regular expression (\code{\link[base]{regex}}) matching.
Names and titles of the matched help entries are displayed nicely formatted. Vignette names, titles and keywords and demo names and titles may also be searched.
}
\details{
for slow internet connections, a timeout can be set for getting the page of RDocumentation via options("RDocumentation.timeOut" = \code{nb_of_seconds}) the default timeout is 3 seconds
}
\examples{
\dontrun{
help.search("linear models")    # In case you forgot how to fit linear
                                # models
help.search("non-existent topic")

??utils::help  # All the topics matching "help" in the utils package

help.search("print")            # All help pages with topics or title
                                # matching 'print'
help.search(apropos = "print")  # The same

help.search(keyword = "hplot",fields=c("alias"))  # All help pages documenting high-level
                                                  # plots.
file.show(file.path(R.home("doc"), "KEYWORDS"))  # show all keywords

## Help pages with documented topics starting with 'try'.
help.search("\\\\btry", fields = "alias")
}

}
\seealso{
\url{http://www.RDocumentation.org} for the online version of the documentation, \code{\link[RDocumentation]{help}} for finding help on non-vague topics or \code{\link[utils]{help.search}} for 
documentation of the offline help.
}

