% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_package}
\alias{install_package}
\title{Install a package from CRAN, BioConductor, or GitHub}
\usage{
install_package(mypkg, type)
}
\arguments{
\item{mypkg}{the name of the package you want to install}

\item{type}{the type of the package, type 1 means the package comes from CRAN, type 2 packages are from BioConductor, type 3 packages are from GitHub and type 4 packages are by default part of R.}
}
\description{
Install a package from CRAN, BioConductor, or GitHub
}
\examples{
\dontrun{
  install_package("dplyr", 1)
  install_package("RDocumentation", 3)
}

}

