% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDocumentation.R
\name{rdocs_url}
\alias{rdocs_url}
\title{Return the url for RDocumentation}
\usage{
rdocs_url()
}
\value{
The url for Rdocumentation
}
\description{
Return the url for RDocumentation
}
\examples{
rdocs_url()
}

